curiango/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
Flask-Mail==0.9.1
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.32
PyMySQL==1.1.1
python-dotenv==1.0.1
marshmallow==3.21.3
WeasyPrint==62.3
ldap3==2.9.1
pytz==2024.1